refactored BusinessPeriod, too
